Move line-graph component into subdir
